analytic-service/requirements.txt
Django>=5.2,<6
psycopg2-binary>=2.9
gunicorn>=23
djangorestframework==3.16.1
python-dotenv>=1.1
asgiref>=3.8
whitenoise>=6.1
sqlparse>=0.5
tzdata>=2025.2
drf-yasg==1.21.11
drf-spectacular==0.28.0
requests
pytest
pytest-django

